LayoutTests/fast/viewport/viewport-68.html

<head>
    <title>Check that alpha character after number is ignored and space acts as a valid separator.</title>
    <meta name="viewport" content="width=100x initial-scale=1">
    <script>
        function test() {
            if (window.layoutTestController) {
                layoutTestController.dumpAsText();
                layoutTestController.dumpConfigurationForViewport(320, 352);
            }
        }
    </script>
</head>
<body onload="test();">
