LayoutTests/fast/viewport/viewport-68-expected.txt

CONSOLE MESSAGE: line 3: Viewport argument value "100x" for key "width" was truncated to its numeric prefix.
viewport size 320x352 scale 1.000000 with limits [1.000000, 5.000000]
